pyproject.toml
[project]
name = "cybergator"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = ["python-dotenv (>=1.0.1,<2.0.0)", "dash (>=2.18.2,<3.0.0)", "psycopg2 (>=2.9.10,<3.0.0)", "neo4j (>=5.28.1,<6.0.0)", "bcrypt (>=4.2.1,<5.0.0)", "flask (>=3.0,<4.0)", "flask-cors (>=5.0.1,<6.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


